Email flow improvements
Let's simplify how to deal with emails when not using AWS, by default,
we now have a dummy implementation that logs outgoing emails.

This removes the AWS SES requirement for trying the project.

Also, do minor refactors to the flows that send emails without the background job,
clarifying with comments that this is done because these emails include user tokens
which should not touch our database.