Deploy preview environments just like Vercel does
A new preview environment will be created when a Pull Request is
submitted.

The workflow will comment on the PR with a link to try the PR changes.

There is a limit in the number of preview environments that will run
concurrently, when such a limit is exceeded, the oldest environment
will be destroyed.

A preview environment is composed from all the apps, backend, web, admin,
including a clean postgres database.

The preview environment name is taken from the PR branch.